db setup and connected

+++ internal/types/types.go
@@ -1,9 +1,9 @@
 package types
 
 type Student struct {
-	Id    int    `json:"id"`
+	Id    int64    `json:"id"`
 	Name  string `json:"name" validate:"required"`
-	Age   int    `json:"age" validate:"required"`
+	Age   uint8    `json:"age" validate:"required"`
 	Class string `json:"class" validate:"required"`
 	Email string `json:"email" validate:"required"`
 }
